+++ packages/docs/src/stories/Dialog.stories.tsx
@@ -1,48 +1,48 @@
-import { Button, useDisclosure } from '@nextui-org/react'
-import { Dialog } from '@asyncui/react'
+import { Button, useDisclosure } from "@nextui-org/react";
+import { Dialog } from "@asyncui/react";
 
 export default {
-  title: 'Components/Dialog',
+  title: "Components/Dialog",
   component: Dialog,
   args: {
-    size: 'md',
+    size: "md",
     hideCloseButton: false,
-    backdrop: 'opaque',
+    backdrop: "opaque",
   },
-  tags: ['autodocs'],
+  tags: ["autodocs"],
   argTypes: {
     size: {
-      control: { type: 'select' },
+      control: { type: "select" },
       options: [
-        'xs',
-        'sm',
-        'md',
-        'lg',
-        'xl',
-        '2xl',
-        '3xl',
-        '4xl',
-        '5xl',
-        'full',
+        "xs",
+        "sm",
+        "md",
+        "lg",
+        "xl",
+        "2xl",
+        "3xl",
+        "4xl",
+        "5xl",
+        "full",
       ],
-      description: 'Tamanho do modal',
+      description: "Tamanho do modal",
       table: {
-        defaultValue: { summary: 'md' },
+        defaultValue: { summary: "md" },
       },
     },
     hideCloseButton: {
-      control: { type: 'boolean' },
+      control: { type: "boolean" },
     },
     backdrop: {
-      control: { type: 'inline-radio' },
-      options: ['transparent', 'opaque', 'blur'],
+      control: { type: "inline-radio" },
+      options: ["transparent", "opaque", "blur"],
     },
   },
-  parameters: { layout: 'centered' },
-}
+  parameters: { layout: "centered" },
+};
 
 export const Dialog_ = () => {
-  const { isOpen, onOpen, onOpenChange } = useDisclosure()
+  const { isOpen, onOpen, onOpenChange } = useDisclosure();
 
   return (
     <>
@@ -58,18 +58,18 @@ export const Dialog_ = () => {
             amet consectetur adipisicing elitLorem ipsum dolor sit amet
             consectetur adipisicing elitLorem ipsum dolor sit amet consectetur
             adipisicing elitLorem ipsum dolor sit amet consectetur adipisicing
-            elit.{' '}
+            elit.{" "}
           </p>
           <p className="m-4">
-            Lorem ipsum dolor sit amet consectetur adipisicing elit.{' '}
+            Lorem ipsum dolor sit amet consectetur adipisicing elit.{" "}
           </p>
         </Dialog.Content>
         <Dialog.Footer>
-          <button className="flex items-center gap-1 body-semi justify-center rounded px-4 py-3  text-colors-primary-600  border border-solid ">
+          <button className="flex items-center gap-1 body-semi justify-center rounded px-4 py-3 bg-white  text-colors-primary-600  border border-solid ">
             Adicionar Conteúdo
           </button>
         </Dialog.Footer>
       </Dialog.Root>
     </>
-  )
-}
+  );
+};