+++ packages/react/src/components/Dialog/DialogContent.tsx
@@ -1,12 +1,12 @@
 import {
   ModalBody as DialogBodyNextUI,
   ModalBodyProps,
-} from '@nextui-org/react'
-import { ReactNode } from 'react'
-import { twMerge } from 'tailwind-merge'
+} from "@nextui-org/react";
+import { ReactNode } from "react";
+import { twMerge } from "tailwind-merge";
 
 export interface DialogContentProps extends ModalBodyProps {
-  children: ReactNode
+  children: ReactNode;
 }
 
 export function DialogContent({
@@ -15,8 +15,11 @@ export function DialogContent({
   ...props
 }: DialogContentProps) {
   return (
-    <DialogBodyNextUI className={twMerge('gap-3', className)} {...props}>
+    <DialogBodyNextUI
+      className={twMerge("gap-3 bg-white", className)}
+      {...props}
+    >
       {children}
     </DialogBodyNextUI>
-  )
+  );
 }