+++ packages/react/src/components/ModalAlert/Modal.tsx
@@ -2,12 +2,12 @@ import {
   ModalContent,
   Modal as ModalNextUI,
   ModalProps as ModalPropsNextUI,
-} from '@nextui-org/react'
-import { ReactNode } from 'react'
-import { twMerge } from 'tailwind-merge'
+} from "@nextui-org/react";
+import { ReactNode } from "react";
+import { twMerge } from "tailwind-merge";
 
 export interface ModalProps extends ModalPropsNextUI {
-  children: ReactNode
+  children: ReactNode;
 }
 
 export function ModalComponent({
@@ -21,10 +21,10 @@ export function ModalComponent({
     <ModalNextUI
       {...props}
       hideCloseButton={hideCloseButton ?? true}
-      size={size ?? 'lg'}
-      className={twMerge('rounded', className)}
+      size={size ?? "lg"}
+      className={twMerge("rounded  border border-solid", className)}
     >
       <ModalContent>{children}</ModalContent>
     </ModalNextUI>
-  )
+  );
 }